requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
requests==2.31.0

# Video Processing
moviepy==1.0.3
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# Image Processing & AI
rembg==2.0.56
torch==2.1.2
torchvision==0.16.2

# AI/ML APIs
openai==1.6.1
anthropic==0.8.1

# Text-to-Speech
elevenlabs==0.2.27
gTTS==2.4.0
pyttsx3==2.90
azure-cognitiveservices-speech==1.34.0
google-cloud-texttospeech==2.16.1

# QR Code Generation
qrcode[pil]==7.4.2

# Audio Processing
pydub==0.25.1

# Configuration
PyYAML==6.0.1

# Utilities
tqdm==4.66.1
colorama==0.4.6
